backend/requirements.txt
fastapi
uvicorn
python-binance
sqlalchemy
psycopg2-binary

